# Fix redundant predicate reordering