@julianmenzler 2022-04-22
Shall I run benchmark_all.sh?