@Bouncner 2022-04-22
> Shall I run benchmark_all.sh?

Could you elaborate a little bit (1-2 sentences) what the current approach does and what is changed by this PR?
I am haven't looked at this rule for ages. If we cannot rule out that there are cases where LQPs could theoretically change with this PR, we should require the full benchmark suite. If we can rule it out, there is probably no need to run it.