@mweisgut 2022-04-25
I have only looked at it very briefly so far: Having both the new `visited_nodes` map and the `visited_nodes` map in `void visit_lqp(...)` might be confusing here. Some more code documentation and/or name changes could provide some clarity here. In addition, visit_lqp is designed to track which nodes have already been visited. Would it be possible to use a nested visit_lqp only for a predicate chain?