@julianmenzler 2022-04-25
> Having both the new visited_nodes map and the visited_nodes map in void visit_lqp(...) might be confusing here. 

We already have something similar in the `BetweenCompositionRule` or the `SubqueryToJoinRule`.

> visit_lqp is designed to track which nodes have already been visited. Would it be possible to use a nested visit_lqp only for a predicate chain?

The inner `visit_lqp` cannot propagate its visited nodes to an outer `visit_lqp` call, so the answer is no.

> Would it be possible to use a nested visit_lqp only for a predicate chain?

Yes, but I do not see any good reason for this.